+++ include/ImprovedGroupView.hpp
@@ -4,7 +4,9 @@
 
 namespace tinker::api::improved_group_view {
 
-    struct UpdateGroupView final : geode::Event<UpdateGroupView, bool()> {};
+    struct UpdateGroupView final : geode::Event<UpdateGroupView, bool()> {
+        using Event::Event;
+    };
 
     inline void updateGroupView() {
         UpdateGroupView().send();