+++ src/modules/UIScaling.hpp
@@ -8,7 +8,7 @@ class $editorModule(UIScaling) {
     bool onSettingChanged(std::string_view key, const matjson::Value& value) override;
     void onEditor() override;
     void onEditorPauseLayer(EditorPauseLayer* editorPauseLayer) override;
-    void setScaling(float scale, bool toolbar, bool fullReload);
+    void setScaling(float scale, bool toolbar, bool topAlign, bool fullReload);
     void setPauseScaling(float scale);
 
     static float getUIScale();